+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## Added
+- A feature to show a preview of a video on hover over
+- `options.interface.addons.hoverPreview` Hover Preview option. `default: false`
+
 ## [2.3.1] - 2019-09-24
 
 ### Fixed